Add files via upload
This dashboard shows the IT industry financial year 2020 report that shows the actual spend compared to the budget and forecast in different regions, areas of IT like R&D, BU .It highlights the area where there was more spend then the budget and forecast and their actual performance. I have also added a decomposition chart which gives a clear insight about the budget vs forecast of various cost element. IT department, country, region, IT area wise.